working on redefining BoundingBox and Octree
