+++ include/Octree.h
@@ -1,21 +1,22 @@
-#ifndef TREE_H
-#define TREE_H
+#ifndef OCTREE_H
+#define OCTREE_H
 #include "BoundingBox.h"
 #include <stack>
 
 static constexpr double BOX_EPS = 1e-5; // a small nudge given to each dimension of the BoundingBox
-class Tree{
+class Octree{
     public:
-    Tree(): _root(nullptr) {}
-    Tree(const Tree&) = delete;
-    Tree(Tree&&) = default;
-    virtual ~Tree() = default;
+    Octree(): _root(nullptr) {}
+    explicit Octree(NodeList& nodes);
+    Octree(const Octree&) = delete;
+    Octree(Octree&&) = default;
+    virtual ~Octree() = default;
     
-    Tree& operator=(const Tree&) = delete;
-    Tree& operator=(Tree&&) = default;
+    Octree& operator=(const Octree&) = delete;
+    Octree& operator=(Octree&&) = default;
 
-    virtual void insert(Node& node) = 0;
-    virtual NodeList remove(Node& node) = 0;
+    void insert(Node& node);
+    NodeList remove(Node& node);
     /**
      * Inputs:
      *  pos: the position of the particle, must be on the surface of a leaf node
@@ -26,7 +27,7 @@ class Tree{
      *  s is the modified distance that the Point has to travel to reach the next Shape
      *  a raw pointer to Shape (from Node::get()) where the particle lands is returned
     **/
-    virtual Shape* nextNode(const Point& pos, const Direction& dir, Shape* current, double& s) const = 0;
+    Shape* nextNode(const Point& pos, const Direction& dir, Shape* current, double& s) const;
 
     double xMin() const noexcept;
     double xMax() const noexcept;
@@ -36,16 +37,10 @@ class Tree{
     double zMax() const noexcept;
 
     explicit operator bool() const noexcept{ return bool(_root); } // to check if tree is empty
-    friend std::ostream& operator<<(std::ostream& os, const Tree& tree);
-
-    using iterator = NodeList::iterator;
-    using const_iterator = NodeList::const_iterator;
-    using BoxStack = std::stack<BoundingBox*>;
+    friend std::ostream& operator<<(std::ostream& os, const Octree& tree);
 
     protected:
     Node _root;
 };
 
-inline decltype(auto) findVertices(Tree::const_iterator cbegin, Tree::const_iterator cend);
-
-#endif // TREE_H
+#endif // OCTREE_H
